organizing our code
