+++ portfolio/src/index.js
@@ -1,21 +1,20 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import AboutMe from './AboutMe';
-import Experience from './Experience'
-import reportWebVitals from './reportWebVitals';
-import Leadership from './Leadership';
-import Projects from './Projects';
-import Contact from './Contact'
+import './styles/index.css';
+import AboutMe from './sections/AboutMe';
+import Experience from './sections/Experience'
+import Leadership from './sections/Leadership';
+import Projects from './sections/Projects';
+import Contact from './sections/Contact'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <AboutMe />
-    <Experience/>
-    <Leadership/>
-    <Projects/>
-    <Contact/>
+    <Experience />
+    <Leadership />
+    <Projects />
+    <Contact />
   </React.StrictMode>
 );
