--- portfolio/src/projects.json
@@ -1,14 +1,0 @@
-[
-    {
-        "id":1,
-        "title":"RPG",
-        "desc":[],
-        "img":"images/rpg.png"
-    },
-    {
-        "id":2,
-        "title":"CLI",
-        "desc":[],
-        "img":"images/mycli.png"
-    }
-]
